feat: Add destructuring assignment for easier access to media field in given array
Field media in receiving array of articles from API is array of objects. URL for images I need is objects in array that itself is a field of another object.
Example: media: [{[{URL}]}]. It pretty hard to access it so I use destructuring assignment and Optonal(file in Utils) to resolve this problem.
Option is used for checking fields in media - nullable or not they are.